+++ frontend_tests/node_tests/dispatch.js
@@ -384,6 +384,9 @@ run_test("realm settings", ({override}) => {
     event = event_fixtures.realm__update__invite_to_realm_policy;
     test_realm_integer(event, "realm_invite_to_realm_policy");
 
+    event = event_fixtures.realm__update__want_advertise_in_communities_directory;
+    test_realm_boolean(event, "realm_want_advertise_in_communities_directory");
+
     event = event_fixtures.realm__update__name;
 
     test_electron_dispatch(event, (key, val) => {