+++ static/js/server_events_dispatch.js
@@ -224,6 +224,7 @@ export function dispatch_normal_event(event) {
                 video_chat_provider: compose.update_video_chat_button_display,
                 giphy_rating: giphy.update_giphy_rating,
                 waiting_period_threshold: noop,
+                want_advertise_in_communities_directory: noop,
                 wildcard_mention_policy: noop,
             };
             switch (event.op) {